extensions.py: remove import of spack.cmd
